Small change in the structure of the public_settings NextJS setup
- extract the default values and the type in a file which can be
  imported by both the client and server code
- extract the server-only code in a separate file and use the
  server-only package to prevent accidental usages on the client
