+++ front_end/package.json
@@ -89,6 +89,7 @@
     "postcss": "^8",
     "prettier": "3.3.1",
     "prettier-plugin-tailwindcss": "0.5.14",
+    "server-only": "^0.0.1",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   }
