+++ front_end/src/utils/public_settings.server.ts
@@ -1,32 +1,5 @@
-export interface PublicSettings {
-  PUBLIC_MINIMAL_UI: boolean;
-  PUBLIC_TURNSTILE_SITE_KEY: string;
-  PUBLIC_APP_URL: string;
-  PUBLIC_API_BASE_URL: string;
-  PUBLIC_POSTHOG_KEY: string;
-  PUBLIC_POSTHOG_BASE_URL: string;
-  PUBLIC_SENTRY_DSN: string;
-  PUBLIC_GOOGLE_MEASUREMENT_ID: string;
-  PUBLIC_DISALLOW_ALL_BOTS: boolean;
-  PUBLIC_LANDING_PAGE_URL: string;
-  PUBLIC_ALLOW_TUTORIAL: boolean;
-  PUBLIC_ALLOW_SIGNUP: boolean;
-}
-
-export const defaultPublicSettingsValues: PublicSettings = {
-  PUBLIC_MINIMAL_UI: false,
-  PUBLIC_TURNSTILE_SITE_KEY: "",
-  PUBLIC_APP_URL: "http://localhost:3000",
-  PUBLIC_API_BASE_URL: "http://localhost:8000",
-  PUBLIC_POSTHOG_KEY: "",
-  PUBLIC_POSTHOG_BASE_URL: "https://us.i.posthog.com",
-  PUBLIC_SENTRY_DSN: "",
-  PUBLIC_GOOGLE_MEASUREMENT_ID: "",
-  PUBLIC_DISALLOW_ALL_BOTS: true,
-  PUBLIC_LANDING_PAGE_URL: "/",
-  PUBLIC_ALLOW_TUTORIAL: true,
-  PUBLIC_ALLOW_SIGNUP: true,
-};
+import "server-only";
+import { defaultPublicSettingsValues } from "./public_settings";
 
 export function getPublicSettings() {
   const defVals = defaultPublicSettingsValues;
